requirements.txt
# VidyAI / Parvidya — Production Dependencies
# Stack: Flask + MongoDB + LangChain + OpenAI GPT-4o + Image-Diff OCR
# Updated: 2026-03-19

# ── Web Framework ──────────────────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
gunicorn>=22.0.0

# ── OpenAI SDK (GPT-4o / GPT-4o-mini + Whisper + Vision) ─────────────────────
openai>=1.35.0

# ── LangChain Core ────────────────────────────────────────────────────────────
# Pin 0.3.x: 1.x reorganized packages and often breaks `langchain.chains` imports.
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0   # PyPDFLoader (text extraction from digital PDFs)

# ── PDF Processing ────────────────────────────────────────────────────────────
pypdf>=4.0.0                    # Extract text layer from digital PDFs
python-docx>=1.1.0             # Parse .docx Word documents (text extraction)
pymupdf>=1.24.0                 # PDF → PIL images for OCR (no system deps — works on Vercel)
pdf2image>=1.17.0               # Local fallback PDF converter (requires poppler locally)
                                # IMPORTANT: pdf2image requires system package 'poppler-utils':
                                #   Ubuntu/Debian: apt-get install -y poppler-utils
                                #   macOS:         brew install poppler
                                #   Alpine:        apk add poppler-utils
                                #   Vercel: NOT supported — use pymupdf instead (installed above)

# ── Image Processing (handwriting OCR pipeline) ───────────────────────────────
Pillow>=10.0.0                  # PIL — image manipulation, contrast enhancement
numpy>=1.26.0                   # Pixel-diff array operations for handwriting isolation
opencv-python-headless>=4.10.0  # Perspective correction, deskew helpers, CLAHE
deskew>=1.5.1                   # Lightweight skew-angle estimation for scanned sheets

# ── PDF Generation (blank answer sheet for image-diff) ───────────────────────
reportlab>=4.0.0                # Generate blank answer sheet PDFs from exam data

# ── Google Cloud Vision (handwriting OCR for subjective answers) ───────────────
google-cloud-vision>=3.7.0

# ── MongoDB — Primary Database ────────────────────────────────────────────────
pymongo>=4.6.0                  # SRV support is built into modern pymongo
dnspython>=2.6.0                # Required by pymongo SRV
certifi>=2024.2.2               # TLS cert bundle for Atlas SSL

# ── Environment / Config ──────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Tokenisation (LangChain token counting) ───────────────────────────────────
tiktoken>=0.7.0

# ── Async Task Queue (optional — graceful degradation if Redis unavailable) ───
celery>=5.3.0                   # Async evaluation tasks (bulk / multi-student)
redis>=5.0.0                    # Celery broker + result backend
                                # Requires Redis server: REDIS_URL env var
                                # Falls back to synchronous evaluation if unavailable

# ── Email ─────────────────────────────────────────────────────────────────────
# Uses stdlib smtplib — no extra package needed

# ── Payment Gateway ───────────────────────────────────────────────────────────
razorpay>=1.4.0                 # Razorpay Python SDK (UPI, cards, net banking, EMI)

# ══════════════════════════════════════════════════════════════════════════════
#  DEPLOYMENT NOTES
# ══════════════════════════════════════════════════════════════════════════════

# ── Local Development ─────────────────────────────────────────────────────────
# pip install -r requirements.txt
# Install poppler (required by pdf2image):
#   macOS:          brew install poppler
#   Ubuntu/Debian:  sudo apt-get install -y poppler-utils
# Optional Redis (for async evaluation):
#   macOS:          brew install redis && brew services start redis
#   Ubuntu/Debian:  sudo apt-get install -y redis-server

# ── Docker ────────────────────────────────────────────────────────────────────
# Add to Dockerfile:
#   RUN apt-get update && apt-get install -y poppler-utils && rm -rf /var/lib/apt/lists/*

# ── Vercel (Python Serverless) ────────────────────────────────────────────────
# 1. Set env vars in Vercel dashboard:
#      OPENAI_API_KEY, MONGO_URI, SMTP_HOST, SMTP_USER, SMTP_PASS, SMTP_PORT
#      REDIS_URL (optional — for async bulk evaluation)
# 2. /tmp is the only writable directory (IS_VERCEL=1 auto-detects this).
# 3. poppler-utils: Vercel Lambda does NOT support apt-get.
#    For Vercel deployment, pdf2image/poppler won't work.
#    The app gracefully falls back to GPT-4o vision-only OCR (Path C).
#    For full image-diff OCR support, use Railway, Render, or Docker instead.
# 4. Build React frontend first:  cd vidyai-react && npm run build
# 5. Point Flask static_folder to vidyai-react/dist/

# ── Railway / Render / Fly.io ─────────────────────────────────────────────────
# These platforms support custom buildpacks. Add a Nixpacks config or Dockerfile:
#   apt-get install -y poppler-utils
# All features including image-diff OCR work fully on these platforms.
